+++ R/bizicount-package.R
@@ -21,7 +21,9 @@
 #'   methods from the \code{\link[texreg]{texreg-package}} and
 #'   \code{\link[DHARMa]{DHARMa}} are extended to support this
 #'   package, namely for the purposes of producing professional tables and
-#'   carrying out post-estimation diagnostics.
+#'   carrying out post-estimation diagnostics. A generic for He et al. (2019)'s
+#'   test for zero-modification is provided, with methods for both `bizicount`
+#'   and `glm`-class objects.
 #'
 #' @section Bivariate Functions:
 #'
@@ -43,6 +45,10 @@
 #' \item \code{\link{simulate.bizicount}} -- Method that simulates observations
 #' using the fitted model's parameters, primarily for use with DHARMa.
 #'
+#' \item \code{\link{zi_test}} -- Method for testing for marginal zero-modification
+#' using the esimated parameters from the model. This test is preferable to the
+#' Vuong, Wald, Score, and LR tests. See He et al. (2019).
+#'
 #'
 #' }
 #'
@@ -60,7 +66,12 @@
 #' \item \code{\link{simulate.zicreg}} -- Method for simulating from the fitted
 #' model. Results are generally used for creating DHARMa objects.
 #'
+#' #' \item \code{\link{zi_test}} -- Method for testing for univariate zero-modification
+#' using the esimated parameters from the model. This test is preferable to the
+#' Vuong, Wald, Score, and LR tests. See He et al. (2019).
+#'
 #' }
+#'
 #'
 #' @author John Niehaus
 #' @docType package