+++ R/zi_dist_fns.R
@@ -1,4 +1,6 @@
 # zi poisson pmf ===============================================================
+na_warn = "NaNs produced; perhaps `lambda` negative, or `psi` not in [0, 1]."
+na_warn_nb = "NaNs produced; perhaps `mu` or `size` negative, or `psi` or `prob` not in [0, 1]?"
 
 #' @name zip
 #' @title The zero-inflated Poisson (ZIP) distribution
@@ -62,44 +64,42 @@
 #' @export
 dzip = function(x, lambda, psi, log = FALSE, recycle = FALSE){
 
-  if(!env_has(e.check, "zi.dens.checks"))
-    check_dist_args(recycle=recycle)
+     if(!env_has(e.check, "zi.dens.checks"))
+          check_dist_args(recycle=recycle)
 
-  call = match.call()
+     call = match.call()
 
-  l = log
-  rm(log)
+     l = log
+     rm(log)
 
-  n = max(length(x), length(lambda), length(psi))
+     n = max(length(x), length(lambda), length(psi))
 
-    if((length(lambda)==1 && lambda < 0) ||
-     (length(psi) ==1 && (psi < 0 || psi > 1))){
-    warning("NaNs produced; perhaps `lambda` negative, or `psi` not on unit-interval?")
-    return(rep(NaN, n))
-  }
+     if( all(lambda < 0) || !any(vprob(psi))){
+          warning(na_warn)
+          return(rep(NaN, n))
+     }
 
-  new.args = lapply(list(psi=psi, lambda=lambda), rep_len.custom, n=n)
-  list2env(new.args, envir=environment())
+     new.args = lapply(list(psi=psi, lambda=lambda), rep_len.custom, n=n)
+     list2env(new.args, envir=environment())
 
-  # note that doing logs this way is computationally superior, as it uses the internal logarithm
-  # for the dpois function and log1p for small psi/1-psi
-  # computing log format first, then exponentiating if necessary also avoids issues with log(0)
-  # because sometimes 0 occurs due to underflow
-  x = rep(x, length.out=n)
-  pmf = alter.cond(
-      ifelse(
-        x==0,
-        log(psi +  exp( log1p(-psi) + dpois(x=x, lambda=lambda, log=T) )),
-        log1p(-psi) + dpois(x=x, lambda=lambda, log=T)),
-      new.message = "NaNs produced; perhaps `lambda` negative, or `psi` not on unit-interval?"
-  )
+     # note that doing logs this way is computationally superior, as it uses the internal logarithm
+     # for the dpois function and log1p for small psi/1-psi
+     # computing log format first, then exponentiating if necessary also avoids issues with log(0)
+     # because sometimes 0 occurs due to underflow
+     x = rep(x, length.out=n)
+     x0 = x == 0
+     pmf = alter.cond(log1p(-psi) + dpois(x = x, lambda = lambda, log = T), suppress = T)
+     pmf[x0] = suppressWarnings(log(psi + exp(pmf))[x0])
 
-  pmf[lambda < 0 | psi < 0 | psi > 1] = NaN
+     pmf[psi < 0 | psi > 1] = NaN
 
-  if(!l)
-    pmf = exp(pmf)
+     if(anyNA(pmf))
+          warning(na_warn)
 
-  return(pmf)
+     if(!l)
+          pmf = exp(pmf)
+
+     return(pmf)
 
 }
 
@@ -108,26 +108,27 @@ dzip = function(x, lambda, psi, log = FALSE, recycle = FALSE){
 #' @export
 rzip = function(n, lambda, psi, recycle = FALSE){
 
-  check_dist_args(recycle=recycle)
+     check_dist_args(recycle=recycle)
 
-  call = match.call()
+     call = match.call()
 
-  if((length(lambda)==1 && lambda < 0) ||
-     (length(psi) ==1 && (psi < 0 || psi > 1))){
-    warning("NaNs produced; perhaps `lambda` negative, or `psi` not on unit-interval?")
-    return(rep(NaN, n))
-  }
+     if( all(lambda < 0) || !any(vprob(psi))){
+          warning(na_warn)
+          return(rep(NaN, n))
+     }
 
-  new.args = lapply(list(psi=psi, lambda=lambda), rep_len.custom, n=n)
-  list2env(new.args, envir=environment())
+     new.args = lapply(list(psi=psi, lambda=lambda), rep_len.custom, n=n)
+     list2env(new.args, envir=environment())
 
-  x = alter.cond(
-    ifelse(rbinom(n,1,psi) == 1, 0, rpois(n, lambda=lambda))
-  )
+     ind = suppressWarnings(rbinom(n, 1, psi))
+     x = suppressWarnings(rpois(n, lambda=lambda))
 
-  x[psi <0 | psi> 1 | lambda < 0] = NaN
+     x[ind == 1 & !is.na(x)] = 0
+     x[is.na(ind) | is.na(x)] = NaN
+     if(anyNA((x)))
+          warning(na_warn)
 
-  return(x)
+     return(x)
 }
 
 
@@ -144,22 +145,23 @@ pzip = function(q, lambda, psi, lower.tail = TRUE, log.p = FALSE, recycle = FALS
   call = match.call()
 
   n = max(length(q), length(lambda), length(psi))
-  if((length(lambda)==1 && lambda < 0) ||
-     (length(psi) ==1 && (psi < 0 || psi > 1))){
-    warning("NaNs produced; perhaps `lambda` negative, or `psi` not on unit-interval?")
-    return(rep(NaN, n))
+  if( all(lambda < 0) || !any(vprob(psi))){
+     warning(na_warn)
+     return(rep(NaN, n))
   }
 
   new.args = lapply(list(psi=psi, lambda=lambda), rep_len.custom, n=n)
   list2env(new.args, envir=environment())
 
-  cdf = alter.cond(
-      psi + (1-psi)*ppois(q, lambda),
-      new.message = "NaNs produced; perhaps `lambda` negative, or `psi` not on unit-interval?"
-  )
+  cdf = suppressWarnings(psi + (1-psi)*ppois(q, lambda))
 
-  cdf[lambda < 0 | psi > 1 | psi <0] = NaN
+
+
   cdf[q < 0] = 0
+  cdf[psi > 1 | psi < 0 | lambda < 0] = NaN
+
+  if(anyNA(cdf))
+       warning(na_warn)
 
   if(!lower.tail)
     cdf = 1 - cdf
@@ -185,8 +187,8 @@ qzip = function(p, lambda, psi, lower.tail = TRUE, log.p = FALSE, recycle = FALS
 
   n = max(length(p), length(lambda), length(psi))
 
-  if(all(lambda < 0) || all(psi < 0 | psi > 1) || all(p < 0)){
-    warning("NaNs produced; perhaps `lambda` negative, or `p`, `psi` not on unit-interval?")
+  if( all(lambda < 0) || !any(vprob(psi)) || !any(vprob(p)) ){
+    warning(na_warn)
     return(rep(NaN, n))
   }
 
@@ -201,16 +203,13 @@ qzip = function(p, lambda, psi, lower.tail = TRUE, log.p = FALSE, recycle = FALS
 
   p2 = (p - psi)/(1 - psi)
 
-  quant= alter.cond(
-    qpois(p2, lambda = lambda),
-    suppress = T
-  )
+  quant = suppressWarnings( qpois(p2, lambda = lambda) )
 
   quant[p < psi] = 0
-  quant[lambda < 0 | psi< 0 | psi > 1 | p< 0 | p>1] = NaN
+  quant[psi < 0 | psi > 1 | p < 0 | p > 1 | lambda < 0] = NaN
 
-  if(any(is.na(quant))){
-    warning("NaNs produced; perhaps `lambda` negative, or `p`, psi` not on unit-interval?")
+  if(anyNA(quant)){
+    warning(na_warn)
   }
 
   return(quant)
@@ -276,50 +275,41 @@ qzip = function(p, lambda, psi, lower.tail = TRUE, log.p = FALSE, recycle = FALS
 #' @export
 dzinb = function(x, size, psi, mu = NULL, prob = NULL, lower.tail = TRUE, log = FALSE, recycle = FALSE){
 
-  if(!env_has(e.check, "zi.dens.checks")){
-    check_dist_args(negbin=T, recycle=recycle)
-  }
+     if(!env_has(e.check, "zi.dens.checks")){
+          check_dist_args(negbin=T, recycle=recycle)
+     }
 
-  call = match.call()
+     call = match.call()
 
-  l = log
-  rm(log)
+     l = log
+     rm(log)
 
-  if(is.null(mu))
-    mu = size*(1-prob)/prob
+     if(is.null(mu))
+          mu = size*(1-prob)/prob
 
-  n = max(length(x), length(size), length(psi), length(prob), length(mu))
-  if((length(size) == 1 && size <= 0) ||
-     (length(psi) == 1 && (psi > 1 | psi < 0)) ||
-     (length(mu) == 1 && mu < 0) ||
-     length(prob) == 1 && (prob > 1 | prob <= 0)
-  ){
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi` or `prob` not on unit-interval?")
-    return(rep(NaN, n))
-  }
+     n = max(length(x), length(size), length(psi), length(prob), length(mu))
+     if (all(size < 0) || !any(vprob(psi)) || (!is.null(mu) && all(mu < 0)) || !any(vprob(prob))) {
+          warning(na_warn_nb)
+          return(rep(NaN, n))
+     }
 
-  x = rep(x, length.out = n)
-  new.args = lapply(list(size=size, psi=psi, mu=mu, prob=prob), rep_len.custom, n=n)
-  list2env(new.args, envir=environment())
+     x = rep(x, length.out = n)
+     new.args = lapply(list(size=size, psi=psi, mu=mu, prob=prob), rep_len.custom, n=n)
+     list2env(new.args, envir=environment())
 
+     x0 = x == 0
+     pmf =   alter.cond(log1p(-psi) + dnbinom(x=x, size=size, mu=mu, log=T), suppress=T)
+     pmf[x0] = suppressWarnings( log(psi + exp(pmf))[x0] )
 
-  pmf = alter.cond(
-        ifelse(
-          x==0,
-          log(psi +  exp( log1p(-psi) + dnbinom(x=x, size=size, mu=mu, log=T) )),
-          log1p(-psi) + dnbinom(x=x, size=size, mu=mu, log=T)
-        )
-  )
+     pmf[psi < 0 | psi > 1 ] = NaN
 
-  if(!l)
-    pmf = exp(pmf)
+     if(!l)
+          pmf = exp(pmf)
 
-  pmf[mu<0|psi < 0 | psi > 1 | prob > 1 | prob <= 0 | size <= 0] = NaN
+     if(anyNA(pmf))
+          warning(na_warn_nb)
 
-  if(anyNA(pmf))
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi`, `prob` not on unit-interval?")
-
-  return(pmf)
+     return(pmf)
 
 }
 
@@ -338,14 +328,11 @@ pzinb = function(q, size, psi,  mu = NULL, prob = NULL, lower.tail = TRUE, log.p
     mu = size*(1-prob)/prob
 
   n = max(length(q), length(size), length(psi), length(prob), length(mu))
-  if((length(size) == 1 && size <= 0) ||
-     (length(psi) == 1 && (psi > 1 | psi < 0)) ||
-     (length(mu) == 1 && mu < 0) ||
-     length(prob) == 1 && (prob > 1 | prob <= 0)
-  ){
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi` or `prob` not on unit-interval?")
-    return(rep(NaN, n))
+  if (all(size < 0) || !any(vprob(psi)) || (!is.null(mu) && all(mu < 0)) || !any(vprob(prob))) {
+       warning(na_warn_nb)
+       return(rep(NaN, n))
   }
+
 
   new.args = lapply(list(size=size, psi=psi, mu=mu, prob=prob), rep_len.custom, n=n)
   list2env(new.args, envir=environment())
@@ -353,11 +340,11 @@ pzinb = function(q, size, psi,  mu = NULL, prob = NULL, lower.tail = TRUE, log.p
   call = match.call()
 
   cdf = alter.cond(
-      psi + (1-psi)*pnbinom(q=q, size=size, mu=mu, log.p=F)
+      psi + (1-psi)*pnbinom(q=q, size=size, mu=mu, log.p=F), suppress=T
   )
 
   cdf[q<0] = 0
-  cdf[mu < 0 | psi < 0 | psi > 1 | prob > 1 | prob <= 0 | size <= 0] = NaN
+  cdf[mu < 0 | psi < 0 | psi > 1 | prob > 1 | prob <= 0 | size < 0] = NaN
 
   if(!lower.tail)
     cdf = 1 - cdf
@@ -366,7 +353,7 @@ pzinb = function(q, size, psi,  mu = NULL, prob = NULL, lower.tail = TRUE, log.p
     cdf = log(cdf)
 
   if(anyNA(cdf))
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi`, `prob` not on unit-interval?")
+    warning(na_warn_nb)
 
   return(cdf)
 
@@ -386,13 +373,9 @@ qzinb = function(p, size, psi, mu = NULL, prob = NULL, lower.tail = TRUE, log.p 
     mu = size*(1-prob)/prob
 
   n = max(length(p), length(size), length(psi), length(prob), length(mu))
-  if(all(size <= 0) ||
-     all(psi > 1 | psi < 0) ||
-     all(mu < 0) ||
-     (all(prob > 1 | prob <= 0) && !is.null(prob))
-  ){
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi` or `prob` not on unit-interval?")
-    return(rep(NaN, n))
+  if (all(size < 0) || !any(vprob(psi)) || (!is.null(mu) && all(mu < 0)) || !any(vprob(prob)) || !any(vprob(p))) {
+       warning(na_warn_nb)
+       return(rep(NaN, n))
   }
 
   new.args = lapply(list(size=size, psi=psi, mu=mu, prob=prob), rep_len.custom, n=n)
@@ -414,10 +397,10 @@ qzinb = function(p, size, psi, mu = NULL, prob = NULL, lower.tail = TRUE, log.p 
   )
 
   quant[p < psi] = 0
-  quant[mu<0| p<0 | p > 1 | psi < 0 | psi > 1 | prob > 1 | prob <= 0 | size <= 0] = NaN
+  quant[mu < 0 | p < 0 | p > 1 | psi < 0 | psi > 1 | prob > 1 | prob <= 0 | size < 0] = NaN
 
   if(anyNA(quant))
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi`, `prob`, `q` not on unit-interval?")
+    warning(na_warn_nb)
 
   return(quant)
 }
@@ -428,33 +411,33 @@ qzinb = function(p, size, psi, mu = NULL, prob = NULL, lower.tail = TRUE, log.p 
 #' @export
 rzinb = function(n, size, psi, mu = NULL, prob = NULL, recycle = FALSE){
 
-  check_dist_args(negbin=T, recycle=recycle)
-  call = match.call()
+     check_dist_args(negbin=T, recycle=recycle)
+     call = match.call()
 
-  if((length(size) == 1 && size <= 0) ||
-     (length(psi) == 1 && (psi > 1 | psi < 0)) ||
-     (length(mu) == 1 && mu < 0) ||
-     length(prob) == 1 && (prob > 1 | prob <= 0)
-  ){
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi` or `prob` not on unit-interval?")
-    return(rep(NaN, n))
-  }
+     if (all(size < 0) || !any(vprob(psi)) || (!is.null(mu) && all(mu < 0)) || !any(vprob(prob))) {
+          cat("here")
+          warning(na_warn_nb)
+          return(rep(NaN, n))
+     }
 
-  if(is.null(mu))
-    mu =  size*(1-prob)/prob
 
-  new.args = lapply(list(size=size, psi=psi, mu=mu, prob=prob), rep_len.custom, n=n)
-  list2env(new.args, envir=environment())
+     if(is.null(mu))
+          mu =  size*(1-prob)/prob
 
-  x = alter.cond(
-      ifelse(rbinom(n,1,psi) == 1, 0, rnbinom(n, size=size, mu=mu))
-  )
+     new.args = lapply(list(size=size, psi=psi, mu=mu, prob=prob), rep_len.custom, n=n)
+     list2env(new.args, envir=environment())
 
-  x[mu < 0 | psi < 0 | psi > 1 | prob > 1 | prob <= 0 | size <= 0] = NaN
-  if(anyNA(x))
-    warning("NaNs produced; perhaps `mu` or `size` negative, or `psi` or `prob` not on unit-interval?")
+     ind = suppressWarnings(rbinom(n, 1, psi) == 1)
+     x = suppressWarnings(rnbinom(n, size = size, mu = mu))
 
-  return(x)
+     x[ind == 1 & !is.na(x)] = 0
+     x[is.na(ind) | is.na(x)] = NaN
+
+     if(anyNA(x))
+          warning(na_warn_nb)
+
+     return(x)
 }
+